Feature/paging move to room 2.3.0-alpha02 to support paging 3

+++ app/src/main/java/dev/leonlatsch/photok/model/database/entity/Photo.kt
@@ -8,5 +8,5 @@ data class Photo(
     val fileName: String,
     val importedAt: Long,
     val type: PhotoType,
-    @PrimaryKey(autoGenerate = true) val id: Long? = null
+    @PrimaryKey(autoGenerate = true) val id: Int? = null
 )
